up to date files

+++ DBMS_files/html/shopping_cart.php
@@ -4,6 +4,7 @@ session_start();
 
 require_once('../PDO_connect.php');
 if(isset ($_POST['recalculate_payment'])){
+	if(!empty($_SESSION['user_id'])){
 	$qty = $_POST['qty'];
 	print_r($qty);
 	$stmt1 = $pdo -> prepare("SELECT distinct isbn from CART_ITEM, SHOPPING_CART WHERE cart_id = ".$_SESSION['cart_id']);
@@ -12,30 +13,66 @@ if(isset ($_POST['recalculate_payment'])){
 	echo "**RESULT BEGINS**";
 	print_r($result);
 	echo "**RESULT ENDS**";
-//	echo "result: ".print_r($result);
+	echo "result: ".print_r($result);
 
 	foreach ($result as $row){
 		echo $row;
-	$stmt2 = $pdo -> prepare("UPDATE CART_ITEM SET quantity = ".$qty[$row]." where cart_id = ".$_SESSION['cart_id']." AND isbn = '".$row."'");
-	$stmt2 -> execute();
+		$stmt2 = $pdo -> prepare("UPDATE CART_ITEM SET quantity = ".$qty[$row]." where cart_id = ".$_SESSION['cart_id']." AND isbn = '".$row."'");
+		$stmt2 -> execute();
 	}
+	}
+else if(!empty($_SESSION['temp_id'])){
+		$qty = $_POST['qty'];
+		print_r($qty);
+		$stmt1 = $pdo -> prepare("SELECT distinct isbn from TEMP_CART_ITEM, TEMP_SHOPPING_CART WHERE cart_id = ".$_SESSION['cart_id']);
+		$stmt1 -> execute();
+		$result = $stmt1 -> fetchall(PDO::FETCH_COLUMN);
+		echo "**RESULT BEGINS**";
+		print_r($result);
+		echo "**RESULT ENDS**";
+		echo "result: ".print_r($result);
+	
+		foreach ($result as $row){
+			echo $row;
+			$stmt2 = $pdo -> prepare("UPDATE TEMP_CART_ITEM SET quantity = ".$qty[$row]." where cart_id = ".$_SESSION['cart_id']." AND isbn = '".$row."'");
+			$stmt2 -> execute();
+		}
+		}
 
 }
 
 if(isset($_GET['delIsbn'])){
-	$del_isbn = $_GET['delIsbn'];
-	$delStmt = $pdo -> prepare('delete from CART_ITEM where isbn = '.'"'.$del_isbn.'"'." and cart_id = ".$_SESSION['cart_id']);
-	$delStmt -> execute();
-	unset($_SESSION['cart_contents'][$del_isbn]);
+	if(!empty($_SESSION['user_id'])){
+		$del_isbn = $_GET['delIsbn'];
+		$delStmt = $pdo -> prepare('delete from CART_ITEM where isbn = '.'"'.$del_isbn.'"'." and cart_id = ".$_SESSION['cart_id']);
+		$delStmt -> execute();
+	}
+else if(!empty($_SESSION['temp_id'])){
+		$del_isbn = $_GET['delIsbn'];
+		$delStmt = $pdo -> prepare('delete from TEMP_CART_ITEM where isbn = '.'"'.$del_isbn.'"'." and cart_id = ".$_SESSION['cart_id']);
+		$delStmt -> execute();
+		}
 }
+if(!empty($_SESSION['user_id'])){
 $stmt = $pdo -> prepare("select BOOK.isbn as ISBN, title as Title, (select name from AUTHOR where BOOK.author_id = id) as Author, 
 (select name from CATEGORY where BOOK.category_id = id) as Category, 
 (select name from PUBLISHER where BOOK.publisher_id = id) as Publisher, BOOK.price as Price, CART_ITEM.quantity as qty from BOOK, SHOPPING_CART, CART_ITEM 
 WHERE CART_ITEM.isbn = BOOK.isbn AND CART_ITEM.cart_id = SHOPPING_CART.id AND user_id = ".$_SESSION['user_id']);
+}
+else if(!empty($_SESSION['temp_id'])){
+	$stmt = $pdo -> prepare("select BOOK.isbn as ISBN, title as Title, (select name from AUTHOR where BOOK.author_id = id) as Author, 
+	(select name from CATEGORY where BOOK.category_id = id) as Category, 
+	(select name from PUBLISHER where BOOK.publisher_id = id) as Publisher, BOOK.price as Price, TEMP_CART_ITEM.quantity as qty from BOOK, TEMP_SHOPPING_CART, TEMP_CART_ITEM 
+	WHERE TEMP_CART_ITEM.isbn = BOOK.isbn AND TEMP_CART_ITEM.cart_id = TEMP_SHOPPING_CART.id AND user_id = ".$_SESSION['temp_id']);
+	}
 $stmt -> execute();
 $result = $stmt->fetchall(PDO::FETCH_ASSOC);
-
-$total_stmt = $pdo -> prepare("SELECT sum(price * quantity) from CART_ITEM, SHOPPING_CART WHERE cart_id = SHOPPING_CART.id AND user_id = ".$_SESSION['user_id']);
+if(!empty($_SESSION['user_id'])){
+	$total_stmt = $pdo -> prepare("SELECT sum(price * quantity) from CART_ITEM, SHOPPING_CART WHERE cart_id = SHOPPING_CART.id AND user_id = ".$_SESSION['user_id']);
+}
+else if(!empty($_SESSION['temp_id'])){
+	$total_stmt = $pdo -> prepare("SELECT sum(price * quantity) from TEMP_CART_ITEM, TEMP_SHOPPING_CART WHERE cart_id = TEMP_SHOPPING_CART.id AND user_id = ".$_SESSION['temp_id']);
+	}
 $total_stmt -> execute();
 $subtotal = $total_stmt->fetch(PDO::FETCH_COLUMN);
